+++ packages/deployer/test/util/provider.ts
@@ -1,6 +1,9 @@
 import { web3Factory } from '@0xproject/dev-utils';
+import { Provider } from '@0xproject/types';
 import * as Web3 from 'web3';
 
-const web3ProviderConfig = { shouldUseInProcessGanache: true };
-const web3Instance = web3Factory.create(web3ProviderConfig);
-export const provider = web3Instance.currentProvider;
+const providerConfigs = { shouldUseInProcessGanache: true };
+const web3Instance = web3Factory.create(providerConfigs);
+const provider: Provider = web3Instance.currentProvider;
+
+export { provider };
