+++ packages/dev-utils/test/blockchain_lifecycle_test.ts
@@ -8,8 +8,8 @@ import { BlockchainLifecycle, web3Factory } from '../src';
 const expect = chai.expect;
 
 describe('BlockchainLifecycle tests', () => {
-    const web3Provider = web3Factory.getRpcProvider({ shouldUseInProcessGanache: true });
-    const web3Wrapper = new Web3Wrapper(web3Provider);
+    const provider = web3Factory.getRpcProvider({ shouldUseInProcessGanache: true });
+    const web3Wrapper = new Web3Wrapper(provider);
     const blockchainLifecycle = new BlockchainLifecycle(web3Wrapper);
     describe('#startAsync/revertAsync', () => {
         it('reverts changes in between', async () => {
