+++ packages/dev-utils/test/rpc_test.ts
@@ -8,8 +8,8 @@ import { web3Factory } from '../src';
 const expect = chai.expect;
 
 describe('RPC tests', () => {
-    const web3Provider = web3Factory.getRpcProvider({ shouldUseInProcessGanache: true });
-    const web3Wrapper = new Web3Wrapper(web3Provider);
+    const provider = web3Factory.getRpcProvider({ shouldUseInProcessGanache: true });
+    const web3Wrapper = new Web3Wrapper(provider);
     describe('#mineBlockAsync', () => {
         it('increases block number when called', async () => {
             const blockNumberBefore = await web3Wrapper.getBlockNumberAsync();
